+++ pennylane_lightning/core/_measurements_base.py
@@ -32,6 +32,7 @@ from pennylane.measurements import (
     StateMeasurement,
     VarianceMP,
 )
+from pennylane.operation import Observable
 from pennylane.ops import SparseHamiltonian, Sum
 from pennylane.tape import QuantumScript
 from pennylane.typing import Result, TensorLike
@@ -70,16 +71,17 @@ class LightningBaseMeasurements(ABC):
         """Returns the simulation data type."""
         return self._qubit_state.dtype
 
-    def _observable_is_sparse(self, obs):
+    @staticmethod
+    def _observable_is_sparse(obs: Observable) -> bool:
         """States if the required observable is sparse.
 
         Args:
-            obs: PennyLane observable to check sparsity.
+            obs (Observable): PennyLane observable to check sparsity.
 
         Returns:
-            True if the measurement process will use the sparse data representation.
+            True if the measurement process only uses the sparse data representation.
         """
-        return isinstance(obs, SparseHamiltonian) or (obs.has_sparse_matrix and not obs.has_matrix)
+        return isinstance(obs, SparseHamiltonian)
 
     @abstractmethod
     def _measurement_dtype(self):
@@ -116,15 +118,14 @@ class LightningBaseMeasurements(ABC):
             Expectation value of the observable
         """
         if self._observable_is_sparse(measurementprocess.obs):
-            # Internally all sparse operators will be treated as a sparse Hermitian operator.
             # We first ensure the CSR sparse representation.
-            CSR_SparseHermitianObs = measurementprocess.obs.sparse_matrix(
+            CSR_SparseHamiltonian = measurementprocess.obs.sparse_matrix(
                 wire_order=list(range(self._qubit_state.num_wires))
             ).tocsr(copy=False)
             return self._measurement_lightning.expval(
-                CSR_SparseHermitianObs.indptr,
-                CSR_SparseHermitianObs.indices,
-                CSR_SparseHermitianObs.data,
+                CSR_SparseHamiltonian.indptr,
+                CSR_SparseHamiltonian.indices,
+                CSR_SparseHamiltonian.data,
             )
 
         if (
@@ -177,13 +178,13 @@ class LightningBaseMeasurements(ABC):
 
         if self._observable_is_sparse(measurementprocess.obs):
             # Ensuring CSR sparse representation.
-            CSR_SparseHermitianObs = measurementprocess.obs.sparse_matrix(
+            CSR_SparseHamiltonian = measurementprocess.obs.sparse_matrix(
                 wire_order=list(range(self._qubit_state.num_wires))
             ).tocsr(copy=False)
             return self._measurement_lightning.var(
-                CSR_SparseHermitianObs.indptr,
-                CSR_SparseHermitianObs.indices,
-                CSR_SparseHermitianObs.data,
+                CSR_SparseHamiltonian.indptr,
+                CSR_SparseHamiltonian.indices,
+                CSR_SparseHamiltonian.data,
             )
 
         if (
@@ -214,7 +215,7 @@ class LightningBaseMeasurements(ABC):
             Callable: function that returns the measurement result
         """
         if isinstance(measurementprocess, StateMeasurement):
-            if isinstance(measurementprocess, ExpectationMP):
+            if isinstance(measurementprocess, ExpectationMP) and measurementprocess.obs is not None:
                 if self._use_mpi:
                     if isinstance(measurementprocess.obs, (qml.Projector)):
                         return self.state_diagonalizing_gates
@@ -226,7 +227,7 @@ class LightningBaseMeasurements(ABC):
             if isinstance(measurementprocess, ProbabilityMP):
                 return self.probs
 
-            if isinstance(measurementprocess, VarianceMP):
+            if isinstance(measurementprocess, VarianceMP) and measurementprocess.obs is not None:
                 if self._use_mpi:
                     if isinstance(measurementprocess.obs, (qml.Projector)):
                         return self.state_diagonalizing_gates
@@ -234,6 +235,7 @@ class LightningBaseMeasurements(ABC):
                     if isinstance(measurementprocess.obs, (qml.Identity, qml.Projector)):
                         return self.state_diagonalizing_gates
                 return self.var
+
             if measurementprocess.obs is None or measurementprocess.obs.has_diagonalizing_gates:
                 return self.state_diagonalizing_gates
 
@@ -316,7 +318,7 @@ class LightningBaseMeasurements(ABC):
                 group[0].obs
             ):
                 raise TypeError(
-                    "ExpectationMP/VarianceMP of sparse operators cannot be computed with samples."
+                    "ExpectationMP/VarianceMP of sparse observables cannot be computed with samples."
                 )
             if isinstance(group[0], VarianceMP) and isinstance(group[0].obs, Sum):
                 raise TypeError("VarianceMP(Sum) cannot be computed with samples.")