+++ pennylane_lightning/core/src/simulators/lightning_kokkos/observables/tests/Test_ObservablesKokkos.cpp
@@ -153,17 +153,17 @@ TEMPLATE_PRODUCT_TEST_CASE("Hamiltonian", "[Observables]", (StateVectorKokkos),
     }
 }
 
-TEMPLATE_PRODUCT_TEST_CASE("SparseHermitianObs", "[Observables]",
+TEMPLATE_PRODUCT_TEST_CASE("SparseHamiltonian", "[Observables]",
                            (StateVectorKokkos), (float, double)) {
     using StateVectorT = TestType;
-    using SparseHermitianObsT = SparseHermitianObs<StateVectorT>;
+    using SparseHamiltonianT = SparseHamiltonian<StateVectorT>;
 
     SECTION("Copy constructibility") {
-        REQUIRE(std::is_copy_constructible_v<SparseHermitianObsT>);
+        REQUIRE(std::is_copy_constructible_v<SparseHamiltonianT>);
     }
 
     SECTION("Move constructibility") {
-        REQUIRE(std::is_move_constructible_v<SparseHermitianObsT>);
+        REQUIRE(std::is_move_constructible_v<SparseHamiltonianT>);
     }
 }