+++ .github/CHANGELOG.md
@@ -6,10 +6,10 @@
 
 ### Improvements
 
-* Added support for sparse `qml.QubitUnitary` gates for `LQubit`, `LGPU`, and `LKokkos` backends.
+* Added support for sparse `qml.QubitUnitary` gates for `lightning.qubit`, `lightning.gpu`, and `lightning.kokkos` backends.
   [#1068](https://github.com/PennyLaneAI/pennylane-lightning/pull/1068).
 
-* Introduced a generalized sparse gate selection system via the `_observable_is_sparse` method in the base measurement class, enabling future expansion for any number of sparse gates.
+* Introduce a generalized sparse gate selection system via the `_observable_is_sparse` method in the base measurement class, enabling future expansion for any number of sparse observables.
   [#1068](https://github.com/PennyLaneAI/pennylane-lightning/pull/1068).
 
 * Optimize the copy of a input state-vector into the LGPU #1071 
@@ -68,6 +68,12 @@
 * `SX` and `C(SX)` gates are natively supported for all lightning devices.
   [(#731)](https://github.com/PennyLaneAI/pennylane-lightning/pull/731)
 
+* Programs transformed by `qml.defer_measurements` can be executed on `lightning.qubit`.
+  [(#1069)](https://github.com/PennyLaneAI/pennylane-lightning/pull/1069)
+
+* `lightning.qubit` supports `ctrl` and `adjoint` with program capture.
+  [(#1069)](https://github.com/PennyLaneAI/pennylane-lightning/pull/1069)
+
 ### Documentation
 
 ### Bug fixes
@@ -101,6 +107,7 @@
 
 This release contains contributions from (in alphabetical order):
 
+Astral Cai,
 Yushao Chen,
 Amintor Dusko,
 Christina Lee,